Creates a Members Sync command (#109)
This allows for the Guild Owner or Bot Owner (D0cR3d) to initiate a sync and log to the database of all users in the server. This WILL take a long time for large servers. There is a wait of 250ms between each logging in order to not DOS the systems.